Declare timestamp as int and cast to int when created

+++ src/Events/EventBean.php
@@ -52,7 +52,7 @@ class EventBean
     /**
      * Error occurred on Timestamp
      *
-     * @var float
+     * @var int
      */
     private $timestamp;
 
@@ -101,7 +101,7 @@ class EventBean
         $this->contexts = array_merge($this->contexts, $contexts);
 
         // Get current Unix timestamp with seconds
-        $this->timestamp = round(microtime(true) * 1000000);
+        $this->timestamp = (int) round(microtime(true) * 1000000);
 
         // Set Parent Transaction
         if ($parent !== null) {
